refactor macOS menu
